boarding-price-ml-service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
psycopg2-binary
pydantic